requirements.txt
numpy==1.20.3
Pillow==7.2.0
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
torch==1.13.1
torchvision==0.14.1
tqdm==4.57.0
